Add documentation and nutils case.
