--- partitioned-heat-conduction/fenics-fenics/Allclean
@@ -1,11 +1,0 @@
-#!/bin/bash
-
-echo "Cleaning logs and output files from previous runs..."
-
-rm -r precice-run/
-
-rm out/* 
-
-rm -f *.log
-
-echo "Cleaning successful!"
