--- partitioned-heat-conduction/fenics-fenics/README.md
@@ -1,3 +1,0 @@
-# Tutorial for a partitioned heat equation using FEniCS
-
-This tutorial is described in the [preCICE wiki](https://github.com/precice/precice/wiki/Tutorial-for-solving-the-heat-equation-in-a-partitioned-fashion-using-FEniCS).
