--- partitioned-heat-conduction/fenics-fenics/precice-adapter-config-D.json
@@ -1,9 +1,0 @@
-{
-  "participant_name": "HeatDirichlet",
-  "config_file_name": "precice-config.xml",
-  "interface": {
-      "coupling_mesh_name": "DirichletNodes",
-      "write_data_name": "Flux",
-      "read_data_name": "Temperature"
-    }
-}
