+++ partitioned-heat-conduction/fenics/.gitignore
@@ -1,0 +1,7 @@
+*.txt
+venv
+.idea
+*.pyc
+out
+*.log
+*events.json
