+++ partitioned-heat-conduction/fenics/precice-adapter-config-D.json
@@ -1,0 +1,9 @@
+{
+  "participant_name": "HeatDirichlet",
+  "config_file_name": "../precice-config.xml",
+  "interface": {
+      "coupling_mesh_name": "DirichletNodes",
+      "write_data_name": "Flux",
+      "read_data_name": "Temperature"
+    }
+}
