+++ partitioned-heat-conduction/fenics/precice-adapter-config-N.json
@@ -1,0 +1,9 @@
+{
+  "participant_name": "HeatNeumann",
+  "config_file_name": "../precice-config.xml",
+  "interface": {
+      "coupling_mesh_name": "NeumannNodes",
+      "write_data_name": "Temperature",
+      "read_data_name": "Flux"
+    }
+}
